Update jruby version and remove rbx in TravisCI
Builds for jruby were failing because the default version used 1.9 mode,
but the `public_suffix` gem required a ruby version >= 2.0. This commit
includes a version of jruby fixed at 9.1.5.0, which uses ruby 2.3.1.

Builds for Rubinius (rbx) were failing as well and can not be fixed at
this point of time, so we decided to stop supporting them.

`ruby-head` and `jruby-head` were added to the "experimental"
builds, since they will always point to the newest versions available.
Builds of these experimental versions are allowed to fail. For more
information see the [TravisCI docs](https://docs.travis-ci.com/user/customizing-the-build/#Rows-that-are-Allowed-to-Fail).