Renames gem from pactas_itero to billwerk
A few years ago, Pactas changed its name to billwerk.
It is now more than time to do this renaming here as well.

A few years ago, Pactas changed its name to billwerk.
It is now more than time to do this renaming here as well.

There will be another release of the old gem with a deprication warning
about the renaming.